net-dns/dnscrypt-proxy: dnscrypt-proxy log rotation needs a hard restart
There is a race condition where copytruncate causes dnscrypt-proxy to end up
with a logfile it is writing to that is not truncated, since it could be in the
middle of a write (dnscrypt-proxy's file descriptor is still to the middle
of the logfile, I believe).  The logfile ends up having a bunch of null data
at the beginning as it continues to write to it.  This has happened multiples
times on different machines for me.  It's much safer to do a hard restart
after a normal rotation.

[sam: Do it in a new revision of the file instead. No extra revbump
because 2.1.16 is being added in the same push.]

Closes: https://bugs.gentoo.org/763495
Part-of: https://github.com/gentoo/gentoo/pull/18383
Closes: https://github.com/gentoo/gentoo/pull/18383
Signed-off-by: Sam James <sam@gentoo.org>